backend/users-service/requirements.txt
Django>=5.0,<5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-yasg==1.21.8
psycopg2-binary==2.9.11
django-cors-headers==4.6.0
django-environ==0.11.2
gunicorn==23.0.0
django-filter
django>=5.0
djangorestframework
drf-yasg
django-cors-headers
setuptools